# library_distributed_server/docs/coordinator/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Coordinator service for distributed transactions in the library management system",
        "title": "Distributed Transaction Coordinator API",
        "termsOfService": "N/A",
        "contact": {
            "name": "API Support",
            "email": "dev@library.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/borrow": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all borrow records for the current site with pagination (Librarian only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Borrowing"
                ],
                "summary": "Get borrow records with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number (0-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term for reader name or borrow ID",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrow records retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied to this site",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve borrow records",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new book borrowing transaction (Librarian only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Borrowing"
                ],
                "summary": "Create borrow transaction",
                "parameters": [
                    {
                        "description": "Borrow request details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateBorrowRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Borrow record created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.PhieuMuon"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied to this site",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create borrow record",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/borrow/detailed": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get borrow records with book and reader details with pagination (enhanced for Flutter app)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Borrowing"
                ],
                "summary": "Get borrow records with details and pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number (0-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term for book title, author, or reader name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detailed borrow records retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/borrow/return/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return a borrowed book and update its status (Librarian only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Borrowing"
                ],
                "summary": "Return borrowed book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Borrow transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Return details (optional)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.ReturnBookRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book returned successfully"
                    },
                    "400": {
                        "description": "Invalid borrow ID format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied to this site",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to return book",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/manager/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get comprehensive statistics across all sites in the distributed system (Manager only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Get system statistics",
                "responses": {
                    "200": {
                        "description": "System statistics retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.SystemStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Manager role required",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve system statistics",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateBorrowRequest": {
            "description": "Request payload for creating a borrow transaction",
            "type": "object",
            "required": [
                "maDG",
                "maQuyenSach"
            ],
            "properties": {
                "maDG": {
                    "description": "Reader ID",
                    "type": "string",
                    "example": "DG001"
                },
                "maQuyenSach": {
                    "description": "Book copy ID",
                    "type": "string",
                    "example": "QS001"
                }
            }
        },
...
        "models.ErrorResponse": {
            "description": "Generic error response",
            "type": "object",
            "properties": {
                "details": {
                    "description": "Error details (optional)"
                },
                "error": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "models.ListResponse": {
            "description": "Generic paginated list response matching Flutter BookListModel structure",
            "type": "object",
            "properties": {
                "items": {
                    "description": "List of items (matches Flutter items field)"
                },
                "paging": {
                    "description": "Pagination info (matches Flutter paging field)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PagingInfo"
                        }
                    ]
                }
            }
        },
        "models.PagingInfo": {
            "description": "Pagination information matching Flutter PagingModel structure",
            "type": "object",
            "properties": {
                "page": {
                    "description": "Current page number (0-based, matches Flutter)",
                    "type": "integer",
                    "example": 0
                },
                "size": {
                    "description": "Items per page (matches Flutter)",
                    "type": "integer",
                    "example": 20
                },
                "totalPages": {
                    "description": "Total number of pages (matches Flutter)",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "models.PhieuMuon": {
            "description": "Borrow transaction (fragmented by branch)",
            "type": "object",
            "required": [
                "maCN",
                "maDG",
                "maQuyenSach"
            ],
            "properties": {
                "maCN": {
                    "description": "Branch code",
                    "type": "string",
                    "example": "Q1"
                },
                "maDG": {
                    "description": "Reader ID",
                    "type": "string",
                    "example": "DG001"
                },
                "maPM": {
                    "description": "Borrow ID (auto-generated)",
                    "type": "integer",
                    "example": 1
                },
                "maQuyenSach": {
                    "description": "Book copy ID",
                    "type": "string",
                    "example": "QS001"
                },
                "ngayMuon": {
                    "description": "Borrow date",
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                },
                "ngayTra": {
                    "description": "Return date (null if not returned)",
                    "type": "string",
                    "example": "2025-01-20T14:00:00Z"
                }
            }
        },
...
        "models.ReturnBookRequest": {
            "description": "Request payload for returning a borrowed book",
            "type": "object",
            "properties": {
                "maQuyenSach": {
                    "description": "Book copy ID (optional)",
                    "type": "string",
                    "example": "QS001"
                },
                "ngayTra": {
                    "description": "Return date",
                    "type": "string",
                    "example": "2025-01-20T14:00:00Z"
                }
            }
        },
...
        "models.SiteStats": {
            "description": "Site-specific statistics",
            "type": "object",
            "properties": {
                "booksOnLoan": {
                    "description": "Books currently on loan",
                    "type": "integer",
                    "example": 75
                },
                "siteID": {
                    "description": "Site identifier",
                    "type": "string",
                    "example": "Q1"
                },
                "totalBooks": {
                    "description": "Total books in this site",
                    "type": "integer",
                    "example": 500
                },
                "totalReaders": {
                    "description": "Total registered readers",
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "models.SystemStats": {
            "description": "Distributed system statistics",
            "type": "object",
            "properties": {
                "statsBySite": {
                    "description": "Statistics by site",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.SiteStats"
                    }
                },
                "totalBooksOnLoan": {
                    "description": "Total books on loan across all sites",
                    "type": "integer",
                    "example": 150
                }
            }
        },
...
    }
}